refs #63 changing table "user" to "users" (Keyword unusable in PG) Removing id=null on insert statement Adding primary key when needed on some create table statement (foreign key not usable)

+++ subsonic-main/src/main/java/net/sourceforge/subsonic/dao/schema/Schema27.java
@@ -35,20 +35,22 @@ public class Schema27 extends Schema{
         if (template.queryForInt("select count(*) from version where version = 3") == 0) {
             LOG.info("Updating database schema to version 3.");
             template.execute("insert into version values (3)");
-
-            LOG.info("Converting database column 'music_file_info.path' to varchar_ignorecase.");
-            template.execute("drop index idx_music_file_info_path");
-            template.execute("alter table music_file_info alter column path varchar_ignorecase not null");
-            template.execute("create index idx_music_file_info_path on music_file_info(path)");
-            LOG.info("Database column 'music_file_info.path' was converted successfully.");
+            //TODO Implement a solution for this other databases
+            if(HSQL_DB == this.getDbType() ){
+	            LOG.info("Converting database column 'music_file_info.path' to varchar_ignorecase.");
+	            template.execute("drop index idx_music_file_info_path");
+	            template.execute("alter table music_file_info alter column path varchar_ignorecase not null");
+	            template.execute("create index idx_music_file_info_path on music_file_info(path)");
+	            LOG.info("Database column 'music_file_info.path' was converted successfully.");
+            }
         }
 
-        if (!columnExists(template, "bytes_streamed", "user")) {
-            LOG.info("Database columns 'user.bytes_streamed/downloaded/uploaded' not found.  Creating them.");
-            template.execute("alter table user add bytes_streamed bigint default 0 not null");
-            template.execute("alter table user add bytes_downloaded bigint default 0 not null");
-            template.execute("alter table user add bytes_uploaded bigint default 0 not null");
-            LOG.info("Database columns 'user.bytes_streamed/downloaded/uploaded' were added successfully.");
+        if (!columnExists(template, "bytes_streamed", "users")) {
+            LOG.info("Database columns 'users.bytes_streamed/downloaded/uploaded' not found.  Creating them.");
+            template.execute("alter table users add bytes_streamed bigint default 0 not null");
+            template.execute("alter table users add bytes_downloaded bigint default 0 not null");
+            template.execute("alter table users add bytes_uploaded bigint default 0 not null");
+            LOG.info("Database columns 'users.bytes_streamed/downloaded/uploaded' were added successfully.");
         }
     }
 }
